Trait clipboard::ClipboardProvider

pub trait ClipboardProvider: Sized {
    fn new() -> Result<Self, Box<dyn Error>>;

    fn get_contents(&mut self) -> Result<String, Box<dyn Error>>;

    fn set_contents(&mut self, _: String) -> Result<(), Box<dyn Error>>;
}

Trait for clipboard access

Required methods

fn new() -> Result<Self, Box<dyn Error>>

Create a context with which to access the clipboard

fn get_contents(&mut self) -> Result<String, Box<dyn Error>>

Method to get the clipboard contents as a String

fn set_contents(&mut self, _: String) -> Result<(), Box<dyn Error>>

Method to set the clipboard contents as a String

Loading content...